src/requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
chardet==4.0.0
comm==0.1.3
contourpy==1.0.7
cramjam==2.6.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
executing==1.2.0
# experiments-utils==2.2.0
fastparquet==2023.2.0
fonttools==4.39.3
fsspec==2023.3.0
idna==2.10
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
joblib==1.2.0
JPype1==1.4.1
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
mlxtend==0.22.0
multiprocess==0.70.13
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.25.1
rulekit==1.6.0
scikit-learn==1.2.2
scipy==1.9.3
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
